Add internal/luksview package
This adds a new package containing a view of the LUKS2 metadata
in a way that is most useful to secboot.

It also contains a representation of tokens that will be used by
secboot. Each token has a name, which is an arbitrary string used to
identify its associated keyslot. The name is intended to be unique
between keyslots and secboot will enforce this. All tokens created by
secboot will have a name and will be associated with one keyslot.

Tokens without a name or which are associated by more than one keyslot
are ignored by the View API.

For context, here's some notes about how the View API is intended to be
used:
- View.TokenByName returns the token for the keyslot with the
supplied name, and is intended to be used when performing changes to
the slot (eg, updating the PCR policy in the TPM case or deleting the
slot).
- View.KeyDataTokensByPriority returns a list of key data tokens
in priority order, and is intended to be used internally by the
ActivateVolumeWithKeyData API. There is no priority for recovery slots
as it doesn't make sense - attempting each slot in a defined order
rather than letting cryptsetup pick the order would require repeated
execution of the KDF if more than one recovery keyslot is defined.
- View.OrphanedTokenIds returns a list of IDs associated with
tokens that have been orphaned (ie, their associated keyslot has been
deleted), and can occur if the process of deleting a keyslot and its
associated token is interrupted. This can be used to resume an
interrupted removal.
- View.UsedKeyslots returns a list of the currently used keyslot IDs.
This will be used by secboot to manually pick an ID for new keyslots
rather than letting cryptsetup handle it, as this makes it easier to
subsequently create a token associated with it.