Bump go-tpm2 version
This bumps the go-tpm2 version to the latest master. This includes
https://github.com/canonical/go-tpm2/commit/96eb110220ece5922dc7b691422fff12735f1880
which makes it possible to decode TPM responses containing TPML_DIGEST_VALUES
structures with SM3 digests.

Some functionality that is not required by the core tpm2 library but was
previously part of the go-tpm2 package has been split out into separate
subpackages.