tpm2: refactor key data validation
SealedKeyObject has a mechanism to perform some consistency
checks of the key data, but this function is huge and has lots
of branches in it to handle the different metadata versions.
Abstract the validation behind an interface that has version
specific implementations to make it a bit more readable.

This also splits validation of the key data and validation of
the auth key into 2 distinct steps, as the latter is only
required before signing PCR policy updates.