tpm2: refactor keyData
SealedKeyObject is just a container around keyData, which
then has several on-disk representations (keyDataRaw_v*). It
also has a mechanism to perform some consistency checks of the
key data (SealedKeyData.validate), but this function is huge
and has branches in it to handle the different metadata versions.

This turns keyData into an interface. Each metadata version
is represented by an implementation of it, with each
implementation containing the version-specific functionality.
Each keyData implementation has its own version-specific
validate function, hopefully making the code a bit more
readable.

This makes the conversion from the on-disk representation
(keyDataRaw_v*) to keyData unnecessary.

This also splits validation of the key data and validation of
the auth key into 2 distinct steps, as the latter is only
required before signing PCR policy updates.

A subsequent PR is going to create an abstraction for some
of the code in tpm2/policy.go, which has similar issues.