tpm2: refactor policy.go
The functions in policy.go are used across all metadata versions, and
some are quite complicated with branches to handle the different
versions and some of them also take a lot of arguments because they
have global scope - this is particularly true of
computeSealedKeyDynamicAuthPolicy (this is in seal.go, but handles
policy stuff), and is a hangover from when we didn't use SealedKeyObject
for updating the PCR policy.

Also in policy.go are some representations of policy data that are
shared between metadata versions, and some version-specific on-disk
representations (similar to the keyData / keyDataRaw_v* split).

This refactors the code in a similar way to the keyData refactor,
such that:
- There is a new interface representing a policy (keyDataPolicy), which
is obtained via the keyData interface.
- The new interface mostly provides a mechanism for updating the PCR
policy and executing the policy.
- The pcrPolicyCounterHandle interface is replaced by
pcrPolicyCounterContext, and can be obtained via the keyDataPolicy
interface.
- Each metadata version has an implementation of keyDataPolicy,
representing the static and PCR policy data.
- The tree structure used for executing PolicyOR assertions has
been split into 2 - there is the existing "flattened" tree
on-disk representation, but this is now converted to an actual
tree.

This does cause a bit of extra code duplication - I've tried to
keep this to a minimum, but I think it makes things more readable
overall.