Add passphrase support to KeyData
This adds passphrase support to KeyData, so that it can be configured to
require a passphrase in order to recover the key. This works by further
encrypting the payload that is protected by the platform's secure device
(such as a TPM) using a passhrase derived key which is derived using the
Argon2i KDF. The encryption is performed with AES-256-CFB.

Platform integration is supported so that passphrase support can take
advantage of features provided by the platform's secure device (such as
dictionary attack protection on TPM devices). The platform implementation
is expected to require the secure device in order to verify whether a
passphrase is correct, both when recovering keys (in order to return an
appropriate error) and when changing the passphrase. This is to prevent
offline attacks against the passphrase. For the TPM platform, this will
be achieved by setting the authorization value of the sealed key object
to a key derived from the passphrase derived key, ensuring that a
compromise of the TPM doesn't reveal the passphrase or passphrase derived
key. The TPM will require knowledge of the authorization value for
unsealing and changing the passphrase.

KeyData makes a distinction between having passphrase support enabled
and having an empty passphrase. There are 3 new APIs for configuring a
passphrase:
- KeyData.SetPassphrase is used to enable passphrase support with the
supplied passphrase if it isn't enabled.
- KeyData.ChangePassphrase is used to update the passphrase.
- KeyData.ClearPassphraseWithPassphrase is used to disable passphrase
support. It is named "ClearPassphraseWithPassphrase" because support for
additional authentication mechanisms in the future could facilitate
clearing the passphrase via another mechanism or just by deleting the
"passphrase_protected_payload" field of KeyData if another authentication
mechanism is enabled (KeyData.ClearPassphrase).

KeyData also provides a new API for recovering keys using a passphrase -
KeyData.RecoverKeysWithPassphrase.