add platform handle accessors to KeyData
This is a prerequisite for creating a tpm2 platform, as we need a way
to create and update a SealedKeyObject from a KeyData object.

There are 2 new functions:
- KeyData.UnmarshalPlatformHandle() - unmarshals the JSON platform
handle payload to the supplied handle.
- KeyData.MarshalAndUpdatePlatformHandle() - marshals the handle to
JSON and updates the in-memory KeyData object.

The intention is for the code in tpm2/ to provide helpers that use
these, rather than being used outside of secboot.

For creating a SealedKeyObject from a KeyData object, there will be
a new API that looks like this:

 func NewSealedKeyObjectFromKeyData(data *secboot.KeyData) (*SealedKeyObject, error)

This will call KeyData.UnmarshalPlatformHandle to obtain the unmarshalled
JSON, which in this case is just a byte slice. It will then call the
existing ReadSealedKeyObject API.

SealedKeyObject will implement json.Marshaler, so you can update the
PCR policy like this:

 sko, _ := NewSealedKeyObjectFromKeyData(keyData)
 sko.UpdatePCRProtectionPolicy(tpm, authKey, profile)
 keyData.MarshalAndUpdatePlatformHandle(sko)
 keyData.WriteAtomic(w)