luks2: add DetectCryptsetupFeatures
We depend on some cryptsetup features that aren't in all of the
cryptsetup versions we have in Ubuntu, so add a simple internal
feature detection API for those. This is motivated by the fact
that we're going to depend on --token-replace, which has just been
approved to land upstream and which we'll need to backport.

Modern versions of cryptsetup have a --test-args option, but token
import and header size settings predate those, so this just uses
version number checks for those.