Salt the digest used for verifying the Snap model HMAC key
KeyData contains a list of HMACs of Snap models that are authorized to
access the data on the device unlocked by it, with the HMAC key being
derived from a hardware protected key that is recovered during early
boot.

In order to verify that the correct HMAC key is supplied when updating
the list of authorized Snap models, we store a digest of the HMAC key.
Update this to store a salted digest of the HMAC key instead, preserving
compatibility with existing KeyData objects.