move Snap model authorization to ActivateVolumeWith*KeyData
The ActivateVolumeWith*KeyData functions currently return a
SnapModelChecker to delegate authorization of the Snap model to
snap-bootstrap. This means that authorization of the model is
performed after activation of the container. There are a couple
of issues with this though:
- If multiple KeyData are supplied and the container is activated
with one of these but the model isn't authorized, it requires another
call to ActivateVolumeWithMultipleKeyData in order to try others.
- If the model isn't auhtorized with any KeyData, the container has
to be deactivated again.

The first issue is difficult to solve with the current architecture
when KeyData are loaded from the LUKS metadata rather than being
supplied externally from a file.

This implements a different approach where SnapModelChecker is removed,
and the model authorization is done in ActivateVolumeWith*KeyData before
the container is activated, with a KeyData being rejected before
activation if the model authorization check fails.