crypt: refactor mocked tests
Mocked tests in crypt_test.go use the internal/luks2 package but mock
the cryptsetup binary. This is a bit strange because the use of cryptsetup
is an implementation detail that's internal to internal/luks2, and is left
over from before that package was created.

Refactor the mocked tests to mock the use of internal/luks2 instead, and
move cryptsetup commandline checks to the internal/luks2 package. This
is going to make it easier to implement tests in future PRs which introduce
APIs that work with tokens.