Allow the platform handle to be any arbitrary JSON value in KeyCreationData
This aligns the representation of the platform handle in KeyCreationData
with the representation introduced to the KeyData API in
https://github.com/snapcore/secboot/pull/191.

Note that the representation of the platform handle in PlatformKeyDataHandler
remains unchanged (it is the encoded JSON). This is becuase
PlatformKeyDataHandler is used to call from KeyData to an implementation,
and it is not possible to decode the JSON in this scenario because the
data is opaque.