crypt: allow keyslots to be named
This updates the APIs in crypt.go that add / modify keys to create
named keyslots - a keyslot with an associated token with the type
"ubuntu-fde" or "ubuntu-fde-recovery" - each of these tokens has
a unique name.

- IntializeLUKS2Container now creates a named keyslot with the
type "ubuntu-fde" and the name "default" unless another name is
supplied.
- AddRecoveryKeyToLUKS2Container has been renamed to
AddLUKS2ContainerRecoveryKey and creates a named recovery keyslot
with the type "ubuntu-fde-recovery" and the name "default-recovery"
unless another name is supplied.
- ChangeLUKS2ContainerUsingRecoveryKey has been renamed to
AddLUKS2ContainerUnlockKey and creates a named keyslot with the type
"ubuntu-fde" and the name "default" unless another name is supplied.
- DeleteLUKS2ContainerKey deletes a named keyslot.
- RenameLUKS2ContainerKey renames a named keyslot.
- ListLUKS2ContainerUnlockKeyNames and ListLUKS2ContainerRecoveryKeyNames
can be used to list the names of existing named keyslots.

A subsequent PR will provide an API for updating the "ubuntu-fde"
tokens in order to attach and update the associated KeyData.