Support activation with KeyData stored in the LUKS header
This adds the ability to activate a LUKS container using KeyData stored in the
LUKS header. It also removes the ActivateVolumeWithMultipleKeyData function -
instead, the ActivateVolumeWithKeyData function accepts a variable number of
externally loaded KeyData objects, such as those loaded from a file. In this
case, the externally provided objects are tried first.