tpm2: require a simulator for createPcrPolicyCounter test
policySuite.TestCreatePcrPolicyCounter uses
TPMTest.NextAvailableHandle() from go-tpm2's testutil package to
pick a free handle for a NV index, but this is inappropriate because
the test requires that the index is created at a specific handle,
and will fail if that handle is already in use and the test picks
a different one. Stop using it, and ensure the test only runs on a
simulator, where we can guarantee that the selected handle is free.