Support measuring extended classic models with a distinct hash
The simplest approach seemed to be to integrate that info as a high
bit in the grade code, for clarity snapd can grow an internal constant
as well to reserve this.

I moved snapd only partially forward because go-tpm2/testutil needs
fixing first, the signature of UserCommonDataDir has changed,
eventually that code will need to consider two possible directories.