tpm2: updates to PCRProtectionProfile API
The PCRProtectionProfile API represents both a profile and a branch within a
profile, which is a bit strange. This splits it into PCRProtectionProfile (to
represent a whole profile) and PCRProtectionProfileBranch (to represent a
single branch within a profile). A profile always has at least a root branch.

The new PCRProtectionProfileBranch API also has a new way of working with
sub-branches, which will hopefully simplify the interaction between it and the
code in efi/, once it is ported to it.

Gone are the immediate panics for some functions when building a profile.
Instead, incorrect use of the API sets an error on the profile which is
returned when trying to compute PCR values.

The new API also only allows a profile to be appended to. The previous API
permitted adding instructions in to the middle of a profile by appending to
a sub-branch that was added via a previous call to AddProfileOR. That is now
explicitly an error because using the API in that way is probably a mistake.

Follow up PRs will port existing code to the new API and make other cleanups.