tpm2: refactor the execution of PCRProtectionProfile
PCRProtectionProfile contains 2 execution loop implementations - one for
String() and one for ComputePCRValues(), each containing fairly messy state
management and making use of type assertions for instruction execution.

Clean this up by implementing a single execution loop that makes use of an
interface for instruction execution and a handler interface supplied by both of
the functions that use it instead.

This enables the opportunity to easily add other methods that might need to
"execute" or loop over a profile, eg, to add a method to serialize a profile
to a binary format in the future and retaining it in the public part of the
key data, which could be useful for debugging purposes.