keydata: use HKDF to derive the snap model HMAC key
The current implementation uses a DRBG where a simpler KDF would be more
appropriate. DRBGs provide some additional features that are just not
required here, which are intended to protect prior and future random
numbers generated from the same source. All of the extra keys derived from
the recovered key payload are all used for protection of the same
resource.

HKDF is a core part of the TLS1.3 key schedule. Using this rather than a
module we've developed is probably going to make future FIPS 140-3 efforts
quite a bit easier given that the go standard library already makes use of
it internally (the version it vendors will most likely require openssl
integration which we could just copy for external use)

This maintains backwards compatibility with exising keys, with a test to
ensure that continues to work.