tpm2: make it possible to serialize / unserialze PCRProtectionProfile
This makes it possible for PCRProtectionProfile to be serialized to
and unserialized from the TPM wire format, using
github.com/canonical/go-tpm2/mu. The profile is serialized in a fairly
space efficient binary format, where digests in the profile are
de-duplicated.

This might be useful in the future in scenarios where it is desirable
to retain a profile, eg, for debugging purposes.