Only use the expand part of HKDF
https://github.com/snapcore/secboot/pull/217 introduced HKDF to derive a
HMAC key. The TPM code is going to use HKDF to derive an elliptic
curve key from the same input key. When working on the latter one and
wondering whether to generate another salt for this, I realized that we
only need to perform the expand step of HKDF. The code I introduced is
currently performing both the extract and expand steps.

From the RFC (https://www.rfc-editor.org/rfc/rfc5869):

> In many applications, the input keying material is not necessarily
> distributed uniformly, and the attacker may have some partial
> knowledge about it (for example, a Diffie-Hellman value computed by a
> key exchange protocol) or even partial control of it (as in some
> entropy-gathering applications).  Thus, the goal of the "extract"
> stage is to "concentrate" the possibly dispersed entropy of the input
> keying material into a short, but cryptographically strong,
> pseudorandom key.  In some applications, the input may already be a
> good pseudorandom key; in these cases, the "extract" stage is not
> necessary, and the "expand" part can be used alone.
>
> The second stage "expands" the pseudorandom key to the desired
> length; the number and lengths of the output keys depend on the
> specific cryptographic algorithms for which the keys are needed.

In our case, the input key is secret and cryptopgraphically strong, which
means that the extract step is unnecessary and we can skip it. This also
means that we don't need an additional salt for the key derivation, as
that's only used in this step.