Change the label used when deriving the policy auth key
Change the label from "PCR-POLICY-AUTH-KEY" to "TPM2-POLICY-AUTH".

The label in keydata.go is "SNAP-MODEL-HMAC" (no "-KEY").

The key derived in the tpm2/ directory is currently used for authorizing
PCR policies but could theoretically be used for authorizing other
policies, should there be a need to add them in the future. In this case,
the TPM2 spec already has a way of limiting the scope of signed policies,
via the policy reference - this is currently set to a value that includes
the string "AUTH-PCR-POLICY".

Setting this to "TPM2-POLICY-AUTH" seems more correct, and consistent
with that in keydata.go as well.