tpm2: Add initial PlatformKeyDataHandler implementation
This adds an initial TPM2 implementation of PlatformKeyDataHandler, plus
new APIs for sealing keys to make use of secboot.KeyData.

The new sealing APIs are ProtectKeyWithTPM, ProtectKeysWithTPM and
ProtectKeyWithExternalStorageKey, to replace the SealKeyToTPM,
SealKeyToTPMMultiple and SealKeyToExternalTPMStorageKey. They are
implemented in a way to facilitate adding another API in the future to
protect additional keys using the same policy as an existing key.

Also added is a new API to update the PCR policy for multiple keys in
one go - UpdateKeyDataPCRProtectionPolicy, which replaces
UpdateKeyPCRProtectionPolicyMultiple.

Obtaining a SealedKeyObject data structure associated with a key is done
differently for keys created with the new APIs. There is a new API to
decode a SealedKeyObject from a secboot.KeyData object - NewSealedKeyData.