keydata: Use the saved key size consistently when using a passphrase
The passphrase derived encryption key size is stored in the key data
when a passphrase is set but is then not used consistently after that.
It's used to calculate the length of key material to derive from the
passphrase, but then the compiled in constant key length is used to
index into the returned bytes in order to separate the encryption key
from the IV. This will cause a panic if the saved key data is
modified to reduce the value of the key size.

This also renames the passphraseDerivdKeyLen constant - it's not the
correct description as the actual derived key length is one block size
larger than this.