Rename KeyCreationData to KeyCreationParams
"Creation data" is a bit of an overloaded term when using a TPM, because
it specifically refers to data that can be used to prove the creation
environment of a TPM object. I'm going to add support to the tpm2
package for saving the creation data to facilitate subsequent
attestation of the state of the initial install environment, so rename
this struct to avoid confusion.

The PR I've proposed to add a new ProtectKeyWithTPM API
(https://github.com/snapcore/secboot/pull/202) also uses
"params" in this context, so it's consistent with that.