efi: updates to ImageLoadEvent
This is the start of a series of changes that significantly refactor the
code in the efi package. These changes will:
- Make the code more modular and maintainable.
- Add support for applying secure boot revocations, including updating
  shim.
- Replace the existing AddSecureBootPolicyProfile and
  AddBootManagerCodeProfile APIs with a single API that will add an
  entire profile. This will also replace the need to use the existing
  AddSystemdStubProfile and tpm2.AddSnapModelProfile APIs.

This last one requires the ability to specify external parameters that
are not part of an executable binary, which is partly what this PR adds.
Note that this does add some APIs that aren't used yet. It also
implements ImageLoadEventSource as one of these new parameters, which
is a bit strange - this is temporary because it's going to be removed
entirely in a future PR, but it needs to stay for now to keep the
existing code working.