efi: add helpers for profiles that make changes to EFI variables
Some profiles may have branches that contain components that make
changes to EFI variables (such as applying SBAT updates) that could
affect other branches in a profile. This introduces a generic mechanism
for facilitating this.

This is based around rootVarsCollector, which collects a set of unique
starting states, each of which is represented by rootVarReader. A starting
state is defined by a set of EFI variable updates applied on top of the
current host environment by a profile. The profile generation code will
contain an outer loop that creates a branch for each of these starting
states. The starting state has a unique cryptographic identifier (using
a SHA1) in order to de-duplicate them and to avoid creating multiple
branches with the same starting state.

For each iteration of the profile generation with an initial starting
state, each branch in the profile will be associated with a
varBranchState - this corresponds to the starting state plus the sequence
of the updates applied by the corresponding branch. Each individual
variable update is tracked to ensure that the profile will incorporate
intermediate states (eg, a SBAT update applied by shim may involve
updating 2 variables and the generated profile has to accomodate that
process being interrupted).

This new functionality is going to be used both for SBAT updates and
signature database updates.