efi: add skeleton for new AddPCRProfile API
This API will replace the existing APIs (AddBootManagerProfile,
AddSecureBootPolicyProfile, AddSystemdStubProfile) in future PRs.

Note that the code in this PR isn't sufficient to actually generate
a profile yet because it's missing implementations of imageLoadHandler
and imageLoadHandlerMap which are what will determine how an image
influences a profile. These will come in a subsequent PR.