luks2: use "aes-cbc-essiv:sha256" on 32bit arm HW
Most hardware should use aes-xts-plain64 for LUKS2 but on 32bit
arm the XTS mode cannot be accelerated by the hardware which
leads to very poor perfomance (less than half compared to cbc
mode). Therefore this commit switches the default on 32bit ARM
to aes-cbc-essiv:sha256.

Note that this is a simple approach, we could run "cryptsetup
benchmark" but that seems over engineered (and easy enough to
switch if we need in the future).