efi: use go:embed for source testdata and recreate data during the test
The tests in efi/ rely on a lot of testdata that is generated out of
band using tools/make-efi-testdata and loaded from files. It can be
burdonsome to update this in order to support additional test cases.

Start embedding source testdata with go:embed, and recreate additional
data from that when running the tests, which will make it easier to
add test cases in the future. This also updates the new tests added
as part of this sequence of PRs to use this new testdata. Existing
tests are left untouched - when this code is removed, some of the
existing test data can be cleaned up.

An advantage of this approach is that we can make use of the fact that
peImageHandle and shimImageHandle are interfaces now, which means that
they can be mocked in tests with real signatures copied from binaries
that we have published. This will allow us to have tests that create
actual production-like digests.