efi: Add primitives to specify rules for constructing imageLoadHandlers
This adds some primitives for building rules which specify how to
construct an imageLoadHandler for an EFI image. Note that this doesn't
actually implement any of those rules yet - that will be in the next PR.

This is implemented in a way that allows some rules to be scoped to a
secure boot namespace. Some examples of where this is useful is that
there are some versions of shim where we can't inspect the built-in
SBAT payloads, but we can infer what they are for specific versions of
shim becasue we know exactly what was published under the Microsoft UEFI
CA namespace. Also, we have a hard requirement on shim >= 15.2, but we
know that Canonical had 2 builds of shim 15 signed by Microsoft (the
original versions shipped with UC20) that included the required patches
from 15.2 to make it work, so we can add quirks for those.

This is all intended to make it easy to adapt to future versions of boot
components that introduce breaking changes or temporary bugs without
having to rewrite lots of code.