keydata.go: add NewKeyDataWithPassphrase API
Under the new design, keyData objects will be differentiated
based on whether they are passphrase protected or not. This
property of new keyData objects will now be immutable.

For that reason passphraseData struct is converted to the new
passphraseParams struct which holds the parameters of the passphrase
and doesn't contain an EncryptedPayload field anymore.

For passphrase protected payloads the key derivation has changed and
now it should work as follows:

- a strong key will be derived from the passphrase

- 3 values will be derived from that strong key given as the IKM
  input to an hkdf
  - an encryption key enc with info PASSPHRASE-ENC
  - an IV with info PASSPHRASE-IV
  - an auth key with info PASSPHRASE-AUTH

The encryption key and IV will be used for the encryption/decryption
of keyData's payload which happens on the platform agnostic API. The
auth key will then need to be used as a way of key validation using
the secure device (i.e the authentication value of the sealed object
for the TPM case).

To support unmarshalling of keys generated with the old format,
the new unmarshalV1KeyPayload() was created under keydata_legacy.go.