keydata.go, keydata_legacy.go: refactor handling of authorized snap models
Under the new design, the mechanism used to restrict access to confidential
keys to a specific set of authorized models will be moved out of the common
keyData implementation and into platform implementations that require it.

As an example of the opposite, UEFI+TPM platforms use measured boot already
to enforce restrictions based on snap models.

AuthorizedSnapModels field is made optional and relevant functions are moved
to keydata_legacy.go.

PrimaryKey is no longer an argument of NewKeyData as it is was relevant for
deriving the authKey which was used for checking authorized snap models. It
is now moved to NewKeyDataScope under bootenv.

SnapModelAuthKey*() functions have been renamed to SnapModelHMACKey*() to
better reflect the legacy HMAC behavior.