tpm2: only create an encrypt session in Connection.init if the EK is valid
Now that operations on sealed objects which require parameter encryption
use a separate session salted with the SRK, only create an encrypt
session in Connection.init if the public area of the EK has the expected
properties.

This also relaxes the EK checks slightly - we can use any key that is a
non-duplicable asymmetric storage key.