efi: convert loadParams struct to a map instead
This avoids the need to add a new member to the loadParams struct for
every new parameter. I would also like implementations of
PCRProfileOption to be able to add a list of root options to the
pcrProfileGenerator which can be used to generate a sequence of
loadParams in pcrProfileGenerator.addPCRProfileBranchForVars from which
multiple branches can be created.

I'm hoping that this will help the work for https://github.com/canonical/secboot/pull/412