efi: PCR4: Add support for agents loaded from flash during OS-present
Some newer laptops are loading endpoint management applications that are
shipped with firmware and measuring these to PCR4 in the OS-present
phase. This currently causes us to produce invalid profiles because
everything after the separator gets dropped from the predicted
measurements.

A future PR will provide support for detecting compatibility for FDE,
with an option for detecting and disallowing these agents. In most cases,
the agents should be disabled (and it's possible to disable Absolute
from userspace on Dell laptops), but we should add support here for any
edge cases where these agents cannot be disabled.