efi: PCR4: Add support for certain firmware applications loaded during OS-present
Some newer laptops are loading endpoint management applications that are
shipped with firmware, and these are being loaded as part of the OS-present
environment with the LoadImage API, resulting in them being measured to PCR4.
This currently causes us to mis-predict values for PCR4 because everything
after the separator (the pre-OS to OS-present transition) is assumed to
belong to the OS and gets dropped from the predicted measurements.

A future PR will provide support for detecting compatibility for FDE,
with an option for detecting and disallowing these firmware applications.
In most cases, they should be disabled (and it's possible to disable
Absolute from userspace on Dell laptops), but we should add support here
for specific, known endpoint management applications for any edge cases where
these agents cannot be disabled.

This adds support for any applications where the firmware volume
filename GUID is well known and maps to "AbsoluteAbtInstaller" or
"AbsoluteComputraceInstaller".