hooks/bootscope: Pass an io.Reader to functions that consume randomness
This passes an io.Reader through APIs between hooks and bootscope (which
uses it for signature generation) in order to make it possible for unit
tests to produce deterministic outputs in some cases where that is useful
(eg, testing the result of marshalling).

This also required the addition of a new helper in internal/testutil to
bypass go's internal randutil.MaybeReadByte, which is designed to
introduce non-determinism even for the same sequence of random bytes.