hooks: Accommodate changes in crypto/ecdsa between go 1.21 and go 1.22
The way that the nonce is generated when creating ECDSA signatures
changed between go 1.21 and 1.22, breaking some of our tests that
rely on hardcoded byte sequences and signatures. Thankfully the new
mechanism still consumes the same amount of entropy so this is fairly
easy to work around, and the new behaviour can still be made
deterministic with our BypassMaybeReadByte hack.

This adds a function that converts a set of input parameters to a
hardcoded signature, with alternate versions compiled for different go
versions.